:sparkles: New `Universal.Attributes.BracketSpacing` sniff
New sniff to enforce a fixed number of spaces on the inside of attribute block brackets.
By default, the sniff expects `0` spaces, but this is configurable via a custom ruleset using the `spaces` property.
Also by default, new lines will not be allowed. This setting can also be toggled by changing the `ignoreNewlines` property in a custom ruleset.

Additionally, when `ignoreNewlines` is set to `true`, the sniff will also check there are no blank lines at the start or end of the attribute block.

Note: the `$spacing` and `$ignoreNewlines` property names are chosen to be in line with commonly used property names as used in PHPCS itself.

This sniff can be used with the default settings to address the following rule from PER-CS:

> 12.1 Basics
> ...
> Attribute names MUST immediately follow the opening attribute block indicator `#[` with no space.
>
> The closing attribute block indicator `]` MUST follow the last character of the attribute name or the closing `)` of its argument list, with no preceding space.

Ref: https://www.php-fig.org/per/coding-style/#121-basics

Includes fixers.
Includes unit tests.
Includes documentation.

Fixes 386